fix(project): align conversation metadata
Extract shared inline metadata primitives for compact time and edited labels, and reuse them in feed identity metadata, project activity cards, and project conversation headers. Reuse the same metadata styling for translation controls so translated/original labels align visually with timestamp metadata.

Move the project conversation header card into a focused component so breadcrumb/menu/title/body/status layout is no longer embedded in the full project conversation page. Keep the conversation action menu pinned to the breadcrumb row while the breadcrumb wraps independently.

Expose createdAt and isEdited for project page activities so cards can show created-time metadata and Edited state without relying on updatedAt, which can reflect non-content updates.

Deploy: agora, api
