+++ services/agora/src/pages/dev/project-page.vue
@@ -477,24 +477,32 @@ const baseActivities = [
     slug: "voices-for-change-share-ideas",
     kind: "conversation",
     isClosed: false,
+    createdAt: new Date("2026-04-03T12:00:00.000Z"),
+    isEdited: false,
     stats: { opinionCount: 62, participantCount: 214, voteCount: 1238 },
   },
   {
     slug: "voices-for-change-rank-priorities",
     kind: "vote",
     isClosed: false,
+    createdAt: new Date("2026-04-10T12:00:00.000Z"),
+    isEdited: false,
     stats: { opinionCount: 18, participantCount: 188, voteCount: 941 },
   },
   {
     slug: "voices-for-change-youth-access",
     kind: "conversation",
     isClosed: true,
+    createdAt: new Date("2026-03-20T12:00:00.000Z"),
+    isEdited: true,
     stats: { opinionCount: 34, participantCount: 126, voteCount: 438 },
   },
   {
     slug: "voices-for-change-final-feedback",
     kind: "vote",
     isClosed: true,
+    createdAt: new Date("2026-03-28T12:00:00.000Z"),
+    isEdited: false,
     stats: { opinionCount: 12, participantCount: 82, voteCount: 224 },
   },
 ] satisfies readonly [
