+++ services/shared/src/types/dto.ts
@@ -424,6 +424,8 @@ const zodProjectPageActivity = z
         slug: zodSlugId,
         kind: z.enum(["conversation", "vote"]),
         isClosed: z.boolean(),
+        createdAt: zodDateTimeFlexible,
+        isEdited: z.boolean(),
         title: zodConversationTitle,
         bodyPlainText: zodConversationBodyPlainTextInput.default(""),
         originalContent: zodProjectPageActivityContentVariant,
